requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
Django==4.1
django-environ==0.9.0
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==14.0.0
flake8==5.0.4
gunicorn==20.1.0
isort==5.10.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
